pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp[tasks]>=3.1.1",
    "python-dotenv>=1.2.2",
    "anthropic>=0.45.0",
    "mcp>=1.2.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
]
